requirements.txt
fastapi
uvicorn
pillow
transformers
python-multipart